% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fertility.R
\docType{data}
\name{Fertility}
\alias{Fertility}
\title{Life expectancy and fertility rates in countries}
\format{A data.frame wherein each row is one country in one
of the three years 1990, 2000, or 2011

- `country`: name of the country
- `lifexp_2011`: life expectancy in 2011
- `lifexp_2000`: similar
- `lifexp_1990`: similar
- `fertilrate_2011`: fertility rate in 2011
- `fertilrate_2000`: similar
- `fertilrate_1990`: similar}
\source{
Life expectancy at birth. (2013, October 14). 
Retrieved from http://data.worldbank.org/indicator/SP.DYN.LE00.IN
}
\usage{
data(Fertility)
}
\description{
Life expectancy (at birth) and fertility rates on a 
country-by-country basis in 1990, 2000, and 2011
}
\keyword{datasets}
